[vs] VoQ Switch objects initialization - Local Port OID mapping to System Ports (#703)
SAI emulation support added for VOQ system ports. The changes include
association of local port id to sytems ports that correspond ot the
local ports. In real SAI, mapping between system port and local port is done
by matching the system ports's <switch_id, core index, core port index> tuple.
In VOQ switches, each port blongs to a switch core and is assigned an index.
This mapping is done in hardware configurations. For VS, the mapping is emulated
via a new file coreportindexmap.ini. This file is made available in /usr/share/sonic/hwsku
in the same way how lanemap.ini is made avaialble. The loading and and parsing of
coreportindexmap.ini is done in similar way as it is done for lanemap.ini.
While emulating configuration system ports, if a system port is found to be a
local port (determined using its switch_id), the local port oid corresponding
system port is retrieved from m_port_list and local port atribute is set in the
system port object. This is used by orchagent (portsorch) for system port initialization.